src/requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
colorama==0.4.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.1.4
django-allauth==0.45.0
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-tinymce==3.3.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.22
mysqlclient @ file:///C:/Users/abdul/OneDrive/Documents/Udemy/JASON_FORMATION_PYTHON/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
oauthlib==3.1.1
Pillow==8.4.0
protobuf==3.14.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.3.0
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
